Fix gcc warning on Solaris --> pid_t and %d don't match.
git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/1.3.x@82048 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/http_main.c
@@ -4294,9 +4294,10 @@ static void standalone_main(int argc, char **argv)
 		const char *pidfile = NULL;
 		pidfile = ap_server_root_relative (pconf, ap_pid_fname);
 		if ( pidfile != NULL && unlink(pidfile) == 0)
-		    ap_log_error(APLOG_MARK, APLOG_NOERRNO|APLOG_INFO, server_conf,
-				 "httpd: removed PID file %s (pid=%d)",
-				 pidfile, getpid());
+		    ap_log_error(APLOG_MARK, APLOG_NOERRNO|APLOG_INFO,
+				 server_conf,
+				 "httpd: removed PID file %s (pid=%ld)",
+				 pidfile, (long)getpid());
 	    }
 
 	    ap_log_error(APLOG_MARK, APLOG_NOERRNO|APLOG_NOTICE, server_conf,
